backend/requirements.txt
asgiref==3.8.1
cryptography==43.0.1
dj-database-url==2.2.0
dj-rest-auth==6.0.0
Django==5.1.1
django-allauth==65.0.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
requests==2.32.3